requirements.txt
chardet==5.2.0
openpyxl==3.1.2
pandas==2.1.2
st-pages==0.4.5
streamlit==1.28.0
streamlit-mermaid==0.2.0
request